requirements.txt
coloredlogs==15.0.1
cuda-toolkit==12.9.1
filelock==3.20.0
flatbuffers==25.9.23
fsspec==2025.10.0
humanfriendly==10.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.3
ml_dtypes==0.5.3
mpmath==1.3.0
networkx==3.5
numpy==2.3.4
nvidia-cuda-runtime==13.0.88
nvidia-cuda-runtime-cu12==12.9.79
onnx==1.19.1
onnxruntime==1.23.2
onnxruntime-gpu==1.23.2
packaging==25.0
pillow==12.0.0
platformdirs==4.5.0
protobuf==6.33.1
pycocotools==2.0.10
pycuda==2025.1.2
pyreadline3==3.5.4
pytools==2025.2.5
setuptools==80.9.0
siphash24==1.8
sympy==1.13.1
tensorrt_cu12==10.14.1.48.post1
tensorrt_cu12_bindings==10.14.1.48.post1
tensorrt_cu12_libs==10.14.1.48.post1
tensorrt_cu13==10.14.1.48.post1
tensorrt_cu13_bindings==10.14.1.48.post1
tensorrt_cu13_libs==10.14.1.48.post1
tensorrt_dispatch_cu12==10.14.1.48.post1
tensorrt_dispatch_cu12_bindings==10.14.1.48.post1
tensorrt_dispatch_cu12_libs==10.14.1.48.post1
tensorrt_lean_cu12==10.14.1.48.post1
tensorrt_lean_cu12_bindings==10.14.1.48.post1
tensorrt_lean_cu12_libs==10.14.1.48.post1
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchvision==0.20.1+cu121
typing_extensions==4.15.0
